+++ packages/core/src/contexts/ScaffoldContext.ts
@@ -6,13 +6,13 @@
 import Context from './Context';
 import { Arguments } from '../types';
 
-export default class ScaffoldContext<
-  T = {
-    action: string;
-    generator: string;
-    dry: boolean;
-  }
-> extends Context<T> {
+export interface ScaffoldArgs {
+  action: string;
+  generator: string;
+  dry: boolean;
+}
+
+export default class ScaffoldContext<T = ScaffoldArgs> extends Context<T> {
   action: string;
 
   generator: string;